# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhotos
  version: "2.0"
  description: |
    API for the project submitted for the WASA course.
    Users can log in, register, upload photos, comment photos, like photos, follow each other and ban each other.
    Github link : [...]
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMsg"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMsg"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMsg"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMsg"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMsg"
  schemas:
    Username:
      type: string
      pattern: '.*'
      description: User's username
      example: "Pollo Palla"
      minLength: 3
      maxLength: 10
...
    ErrorMsg:
      description: Error message
      type: string
      example: "There is an error with the database."
    UserInfo:
      description: User's info
      type: object
      properties:
        username:
          $ref: "#/components/schemas/Username"
        numposts:
          description: Number of photos posted by the user
          type: integer
          min: 0
          example: 10
        numfollowers:
          description: Number of people who follow the user
          type: integer
          min: 0
          example: 10
        numfollowing:
          description: Number of people who the user is following
          type: integer
          min: 0
          example: 10
...
  parameters:
    Username:
      name: username
      in: path
      description: user's username
      required: true
      schema:
        $ref: "#/components/schemas/Username"
...
security:
  - BearerAuth: []
paths:
...
  /users/{username}:
      put:
        operationId: setMyUserName
        tags: ["user actions"]
        summary: Change the user's username
        description: |
          Set a new username for the user if it's not already taken.
        parameters:
          - $ref: "#/components/parameters/Username"
        requestBody:
          description: User's username
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Username"
          required: true
        responses:
          "204":
            description: Username updated successfully
          "400":
            $ref: "#/components/responses/BadRequest"
          "409":
            $ref: "#/components/responses/Conflict"          
          "500":
            $ref: "#/components/responses/InternalServerError"
      get:
        operationId: getUserProfile
        tags: ["user actions"]
        summary: Get the user's profile info
        description: |
          Get the info about the user's profile.
        parameters:
          - $ref: "#/components/parameters/Username"
        responses:
          "201":
            description: User's info successfully obtained
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/UserInfo"
          "400":
            $ref: "#/components/responses/BadRequest"
          "403":
            $ref: "#/components/responses/Forbidden"
          "404":
            $ref: "#/components/responses/NotFound"
          "500":
            $ref: "#/components/responses/InternalServerError"
...
  /users/{username}/follows/{followname}:
    put:
      security: []
      operationId: followUser
      tags: ["follow actions"]
      summary: Follow a user
      description: |
        Follow a user
      parameters:
        - $ref: "#/components/parameters/Username"
        - $ref: "#/components/parameters/Username"
      responses:
        "204":
          description: User followed successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      operationId: unfollowUser
      tags: ["follow actions"]
      summary: Unfollow a user
      description: |
        Unfollow a user
      parameters:
        - $ref: "#/components/parameters/Username"
        - $ref: "#/components/parameters/Username"
      responses:
        "204":
          description: User unfollowed successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{username}/bans/{banname}:
    put:
      operationId: banUser
      tags: ["ban actions"]
      summary: Ban a user
      description: |
        Ban a user
      parameters:
        - $ref: "#/components/parameters/Username"
        - $ref: "#/components/parameters/Username"
      responses:
        "204":
          description: User banned successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      operationId: unbanUser
      tags: ["ban actions"]
      summary: Unban a user
      description: |
        Unban a user
      parameters:
        - $ref: "#/components/parameters/Username"
        - $ref: "#/components/parameters/Username"
      responses:
        "204":
          description: User unbanned successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...